fix(model): do not extend types that do not exist
As Team and User models no longer ship with pubsweet-server by default, since version 11.0.0, their
GraphQL types can only be extended if they are included as components, i.e. @pubsweet/model-user or
@pubsweet/model-team. eLife's team and user models are standalone (not extending PubSweet's
Team/User) anyway, and as such the reliance on existing schema/'extend type' is coincidental in a
way and no longer needed or desired.