feat: use latest PubSweet's saveGraph (#2131)
* feat: use latest PubSweet's saveGraph

* chore: fix broken yarn.lock file

* fix: change require path for db

* fix(model): do not extend types that do not exist

As Team and User models no longer ship with pubsweet-server by default, since version 11.0.0, their
GraphQL types can only be extended if they are included as components, i.e. @pubsweet/model-user or
@pubsweet/model-team. eLife's team and user models are standalone (not extending PubSweet's
Team/User) anyway, and as such the reliance on existing schema/'extend type' is coincidental in a
way and no longer needed or desired.

* fix: also resolve mjs files with webpack

* chore: update to latest pubsweet-server and lock apollo-client

* fix: revert saveGraph to save

* chore: bump pubsweet dependencies

* array of strings as save options on _saveTeams

* fix: improve handling of file saves

* test: mock saveGraph instead

* test: skip and comment test for deleting related entities

* fix: add apollo-cache-inmemory to devDependencies

* test: remove test for saving a user with a given id

This opposite of this is tested behaviour in PubSweet: you should be able to save a model with a provided id. If you don't want this, saveGraph({insertMissing: false}) is an option.

* feat: use simple save() instead of graph saving, wherever possible

* test: use saveGraph when needed

* test: created and updated are returned as dates