2194: Fixed issue with PersonPod components being needlessly re-rendered (#2206)
This was happening  whenever an item was being selected or unselected.

The root cause was inside the PersonPodGrid component's render method, where it was passing
an anonymous method to each PersonPod component that calls the toggleSelection method passed
down from the PeoplePickerLogic component. This led to a prop change being detected to a new
anonymous method being created on PersonPodGrid rerender.

The fix involves adding a shouldComponentUpdate method to check whether relevant props have been modified.

Additionally, the `acceptModal` method passed to PersonInfoModal component doesn't accept a `person` parameter anymore as it wasn't actually needed (and was actually given a click event value).